Update README.md
Programmer, This code was very well thought out and done well. However, with your code I was in no way able to access it as long as the code from line 9 to line 19. This being your facebook login user part. I tried to look through the book and fix it but was not able to do so. When I went through this github and deleted the code through those lines it worked really well. However when I was at the end of the process and on the second page of the app I was not able to navigate anywhere in order to reach the first page. There is no option (button) in your code that would make this doable. So adding that would be very nice for the user that would want to get back to the first page. 

Also, I am not sure why but I was not able to import your profile through my github. it continuously gave me a "fatal" error message and would not allow it to go through. I do not know if this is yours or anyones fault but this is what happens. 

After all that I suggest that only a color change to your app would be nice to meet all of the CRAP Principles. 

Good job on your app and good luck working for Nike.